# apps/api/docs/swagger.yaml
definitions:
...
  cashflow.CreateManualCashflowTransactionRequest:
    properties:
      amount:
        type: string
      date:
        type: string
      description:
        type: string
      note:
        type: string
      tag:
        type: string
      type:
        type: string
      vendor:
        type: string
    type: object
  cashflow.CreateTransactionResponse:
    properties:
      amountCents:
        type: integer
      date:
        type: string
      description:
        type: string
      direction:
        type: string
      id:
        type: string
      ignored:
        type: boolean
      note:
        type: string
      source:
        type: string
      tag:
        type: string
    type: object
  cashflow.CreateTransactionsRequest:
    properties:
      transactions:
        items:
          $ref: '#/definitions/cashflow.CreateManualCashflowTransactionRequest'
        type: array
    type: object
...
  cashflow.TransactionsResponse:
    properties:
      created_count:
        type: integer
      data:
        items:
          $ref: '#/definitions/cashflow.CreateTransactionResponse'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /auth/{provider}/login:
    get:
      description: Redirects to the identity provider's authorization endpoint. Browser-facing;
        not callable from XHR.
      parameters:
      - description: Identity provider slug (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Redirect to the identity provider
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start sign-in
      tags:
      - auth
...
  /cashflow/transactions:
    get:
      consumes:
      - application/json
      description: Filter cashflow transactions with explicit field filters and optional
        fuzzy search over description, note, and tag.
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: 'Sort field: date, description, note, tag, source, amount'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc or desc'
        in: query
        name: sort_order
        type: string
      - description: Fuzzy query over description, note, tag
        in: query
        name: q
        type: string
      - description: Case-insensitive contains filter on description
        in: query
        name: description
        type: string
      - description: Case-insensitive contains filter on note
        in: query
        name: note
        type: string
      - description: Case-insensitive contains filter on source
        in: query
        name: source
        type: string
      - description: 'Direction filter: in or out'
        in: query
        name: direction
        type: string
      - description: Comma-separated tags, e.g. food,travel
        in: query
        name: tags
        type: string
      - description: Only untagged transactions (empty tag)
        in: query
        name: untagged
        type: boolean
      - description: Hide ignored transactions
        in: query
        name: hide_ignored
        type: boolean
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cashflow.TransactionsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search cashflow transactions
      tags:
      - Transactions
...
  /cashflow/transactions/manual:
    post:
      consumes:
      - application/json
      description: Creates one or more manual cashflow transactions for an account.
      parameters:
      - description: Manual cashflow transactions payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cashflow.CreateTransactionsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cashflow.TransactionsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create manual cashflow transactions
      tags:
      - Transactions
...
swagger: "2.0"